+++ src/scenes/components/tracer.h
@@ -27,6 +27,7 @@ void Tracer_UpdateGraphics(Tracer tracer);
  * Draws each point, scaled to fit them all within a rectangle whose top-left
  * corner is at (x, y) and dimensions (width, height).
  */
-void Tracer_Draw(Tracer tracer, int x, int y, float depth, int width, int height);
+void Tracer_Draw(Tracer tracer, float x, float y, float depth, float width,
+		float height);
 
 #endif
